Adds top-users endpoint
